KTOR-6946 headers in the default request are overriding custom headers
