+++ ktor-client/ktor-client-plugins/ktor-client-content-negotiation/common/test/io/ktor/client/plugins/ContentNegotiationTests.kt
@@ -413,31 +413,6 @@ class ContentNegotiationTests {
         }
     }
 
-    @Test
-    fun canOverrideTheDefaultRequestType() = testWithEngine(MockEngine) {
-        config {
-            defaultRequest {
-                contentType(ContentType.Application.Json)
-            }
-            engine {
-                addHandler { request ->
-                    respond(
-                        content = """{"x": 123}""",
-                        headers = headersOf("X-ContentType", request.body.contentType.toString())
-                    )
-                }
-            }
-        }
-        test { client ->
-            val response = client.get("/") {
-                setBody("Hello World!")
-                contentType(ContentType.Application.Xml)
-            }
-
-            assertEquals("application/xml", response.headers["X-ContentType"])
-        }
-    }
-
     object Thing
 
     data class StringWrapper(val value: String)
